+++ src/constants.rs
@@ -1,0 +1,4 @@
+//constants
+pub const OK_RESPONSE: &str = "HTTP/1.1 200 OK\r\nContent-Type: application/json\r\n\r\n";
+pub const NOT_FOUND: &str = "HTTP/1.1 404 NOT FOUND\r\n\r\n";
+pub const INTERNAL_SERVER_ERROR: &str = "HTTP/1.1 500 INTERNAL SERVER ERROR\r\n\r\n";
