+++ src/database.rs
@@ -1,0 +1,21 @@
+use postgres::{ Client, NoTls };
+use postgres::Error as PostgresError;
+use std::net::{ TcpListener, TcpStream };
+use std::io::{ Read, Write };
+use std::env;
+
+//set_database function
+pub fn set_database(db_url: &str) -> Result<(), PostgresError> {
+    //Connect to database
+    let mut client = Client::connect(db_url, NoTls)?;
+
+    //Create table
+    client.batch_execute(
+        "CREATE TABLE IF NOT EXISTS users (
+            id SERIAL PRIMARY KEY,
+            name VARCHAR NOT NULL,
+            email VARCHAR NOT NULL
+        )"
+    )?;
+    Ok(())
+}
