+++ src/main.rs
@@ -4,29 +4,35 @@ use std::net::{ TcpListener, TcpStream };
 use std::io::{ Read, Write };
 use std::env;
 
+mod database;
+mod controllers;
+mod constants;
+mod models;
+mod utils;
+
 #[macro_use]
 extern crate serde_derive;
 
-//Model: USer struct with id, name, email
-#[derive(Serialize, Deserialize)]
-struct User {
-    id: Option<i32>,
-    name: String,
-    email: String,
-}
+// //Model: User struct with id, name, email
+// #[derive(Serialize, Deserialize)]
+// struct User {
+//     id: Option<i32>,
+//     name: String,
+//     email: String,
+// }
 
-//DATABASE_URL
+// //DATABASE_URL
 const DB_URL: &str = env!("DATABASE_URL");
 
-//constants
-const OK_RESPONSE: &str = "HTTP/1.1 200 OK\r\nContent-Type: application/json\r\n\r\n";
-const NOT_FOUND: &str = "HTTP/1.1 404 NOT FOUND\r\n\r\n";
-const INTERNAL_SERVER_ERROR: &str = "HTTP/1.1 500 INTERNAL SERVER ERROR\r\n\r\n";
+// //constants
+// const OK_RESPONSE: &str = "HTTP/1.1 200 OK\r\nContent-Type: application/json\r\n\r\n";
+// const NOT_FOUND: &str = "HTTP/1.1 404 NOT FOUND\r\n\r\n";
+// const INTERNAL_SERVER_ERROR: &str = "HTTP/1.1 500 INTERNAL SERVER ERROR\r\n\r\n";
 
 //main function
 fn main() {
     //Set database
-    if let Err(e) = set_database() {
+    if let Err(e) = database::set_database(DB_URL) {
         println!("Error: {}", e);
         return;
     }
@@ -58,12 +64,12 @@ fn handle_client(mut stream: TcpStream) {
             request.push_str(String::from_utf8_lossy(&buffer[..size]).as_ref());
 
             let (status_line, content) = match &*request {
-                r if r.starts_with("POST /users") => handle_post_request(r),
-                r if r.starts_with("GET /users/") => handle_get_request(r),
-                r if r.starts_with("GET /users") => handle_get_all_request(r),
-                r if r.starts_with("PUT /users/") => handle_put_request(r),
-                r if r.starts_with("DELETE /users/") => handle_delete_request(r),
-                _ => (NOT_FOUND.to_string(), "404 Not Found".to_string()),
+                r if r.starts_with("POST /users") => controllers::handle_post_request(r),
+                r if r.starts_with("GET /users/") => controllers::handle_get_request(r),
+                r if r.starts_with("GET /users") => controllers::handle_get_all_request(r),
+                r if r.starts_with("PUT /users/") => controllers::handle_put_request(r),
+                r if r.starts_with("DELETE /users/") => controllers::handle_delete_request(r),
+                _ => (constants::NOT_FOUND.to_string(), "404 Not Found".to_string()),
             };
 
             stream.write_all(format!("{}{}", status_line, content).as_bytes()).unwrap();
@@ -74,127 +80,127 @@ fn handle_client(mut stream: TcpStream) {
     }
 }
 
-//CONTROLLERS
+// //CONTROLLERS
 
-//handle_post_request function
-fn handle_post_request(request: &str) -> (String, String) {
-    match (get_user_request_body(&request), Client::connect(DB_URL, NoTls)) {
-        (Ok(user), Ok(mut client)) => {
-            client
-                .execute(
-                    "INSERT INTO users (name, email) VALUES ($1, $2)",
-                    &[&user.name, &user.email]
-                )
-                .unwrap();
+// //handle_post_request function
+// fn handle_post_request(request: &str) -> (String, String) {
+//     match (get_user_request_body(&request), Client::connect(DB_URL, NoTls)) {
+//         (Ok(user), Ok(mut client)) => {
+//             client
+//                 .execute(
+//                     "INSERT INTO users (name, email) VALUES ($1, $2)",
+//                     &[&user.name, &user.email]
+//                 )
+//                 .unwrap();
 
-            (OK_RESPONSE.to_string(), "User created".to_string())
-        }
-        _ => (INTERNAL_SERVER_ERROR.to_string(), "Error".to_string()),
-    }
-}
+//             (OK_RESPONSE.to_string(), "User created".to_string())
+//         }
+//         _ => (INTERNAL_SERVER_ERROR.to_string(), "Error".to_string()),
+//     }
+// }
 
-//handle_get_request function
-fn handle_get_request(request: &str) -> (String, String) {
-    match (get_id(&request).parse::<i32>(), Client::connect(DB_URL, NoTls)) {
-        (Ok(id), Ok(mut client)) =>
-            match client.query_one("SELECT * FROM users WHERE id = $1", &[&id]) {
-                Ok(row) => {
-                    let user = User {
-                        id: row.get(0),
-                        name: row.get(1),
-                        email: row.get(2),
-                    };
+// //handle_get_request function
+// fn handle_get_request(request: &str) -> (String, String) {
+//     match (get_id(&request).parse::<i32>(), Client::connect(DB_URL, NoTls)) {
+//         (Ok(id), Ok(mut client)) =>
+//             match client.query_one("SELECT * FROM users WHERE id = $1", &[&id]) {
+//                 Ok(row) => {
+//                     let user = User {
+//                         id: row.get(0),
+//                         name: row.get(1),
+//                         email: row.get(2),
+//                     };
 
-                    (OK_RESPONSE.to_string(), serde_json::to_string(&user).unwrap())
-                }
-                _ => (NOT_FOUND.to_string(), "User not found".to_string()),
-            }
+//                     (OK_RESPONSE.to_string(), serde_json::to_string(&user).unwrap())
+//                 }
+//                 _ => (NOT_FOUND.to_string(), "User not found".to_string()),
+//             }
 
-        _ => (INTERNAL_SERVER_ERROR.to_string(), "Error".to_string()),
-    }
-}
+//         _ => (INTERNAL_SERVER_ERROR.to_string(), "Error".to_string()),
+//     }
+// }
 
-//handle_get_all_request function
-fn handle_get_all_request(request: &str) -> (String, String) {
-    match Client::connect(DB_URL, NoTls) {
-        Ok(mut client) => {
-            let mut users = Vec::new();
+// //handle_get_all_request function
+// fn handle_get_all_request(request: &str) -> (String, String) {
+//     match Client::connect(DB_URL, NoTls) {
+//         Ok(mut client) => {
+//             let mut users = Vec::new();
 
-            for row in client.query("SELECT * FROM users", &[]).unwrap() {
-                users.push(User {
-                    id: row.get(0),
-                    name: row.get(1),
-                    email: row.get(2),
-                });
-            }
+//             for row in client.query("SELECT * FROM users", &[]).unwrap() {
+//                 users.push(User {
+//                     id: row.get(0),
+//                     name: row.get(1),
+//                     email: row.get(2),
+//                 });
+//             }
 
-            (OK_RESPONSE.to_string(), serde_json::to_string(&users).unwrap())
-        }
-        _ => (INTERNAL_SERVER_ERROR.to_string(), "Error".to_string()),
-    }
-}
+//             (OK_RESPONSE.to_string(), serde_json::to_string(&users).unwrap())
+//         }
+//         _ => (INTERNAL_SERVER_ERROR.to_string(), "Error".to_string()),
+//     }
+// }
 
-//handle_put_request function
-fn handle_put_request(request: &str) -> (String, String) {
-    match
-        (
-            get_id(&request).parse::<i32>(),
-            get_user_request_body(&request),
-            Client::connect(DB_URL, NoTls),
-        )
-    {
-        (Ok(id), Ok(user), Ok(mut client)) => {
-            client
-                .execute(
-                    "UPDATE users SET name = $1, email = $2 WHERE id = $3",
-                    &[&user.name, &user.email, &id]
-                )
-                .unwrap();
+// //handle_put_request function
+// fn handle_put_request(request: &str) -> (String, String) {
+//     match
+//         (
+//             get_id(&request).parse::<i32>(),
+//             get_user_request_body(&request),
+//             Client::connect(DB_URL, NoTls),
+//         )
+//     {
+//         (Ok(id), Ok(user), Ok(mut client)) => {
+//             client
+//                 .execute(
+//                     "UPDATE users SET name = $1, email = $2 WHERE id = $3",
+//                     &[&user.name, &user.email, &id]
+//                 )
+//                 .unwrap();
 
-            (OK_RESPONSE.to_string(), "User updated".to_string())
-        }
-        _ => (INTERNAL_SERVER_ERROR.to_string(), "Error".to_string()),
-    }
-}
+//             (OK_RESPONSE.to_string(), "User updated".to_string())
+//         }
+//         _ => (INTERNAL_SERVER_ERROR.to_string(), "Error".to_string()),
+//     }
+// }
 
-//handle_delete_request function
-fn handle_delete_request(request: &str) -> (String, String) {
-    match (get_id(&request).parse::<i32>(), Client::connect(DB_URL, NoTls)) {
-        (Ok(id), Ok(mut client)) => {
-            let rows_affected = client.execute("DELETE FROM users WHERE id = $1", &[&id]).unwrap();
+// //handle_delete_request function
+// fn handle_delete_request(request: &str) -> (String, String) {
+//     match (get_id(&request).parse::<i32>(), Client::connect(DB_URL, NoTls)) {
+//         (Ok(id), Ok(mut client)) => {
+//             let rows_affected = client.execute("DELETE FROM users WHERE id = $1", &[&id]).unwrap();
 
-            if rows_affected == 0 {
-                return (NOT_FOUND.to_string(), "User not found".to_string());
-            }
+//             if rows_affected == 0 {
+//                 return (NOT_FOUND.to_string(), "User not found".to_string());
+//             }
 
-            (OK_RESPONSE.to_string(), "User deleted".to_string())
-        }
-        _ => (INTERNAL_SERVER_ERROR.to_string(), "Error".to_string()),
-    }
-}
+//             (OK_RESPONSE.to_string(), "User deleted".to_string())
+//         }
+//         _ => (INTERNAL_SERVER_ERROR.to_string(), "Error".to_string()),
+//     }
+// }
 
-//set_database function
-fn set_database() -> Result<(), PostgresError> {
-    //Connect to database
-    let mut client = Client::connect(DB_URL, NoTls)?;
+// //set_database function
+// fn set_database() -> Result<(), PostgresError> {
+//     //Connect to database
+//     let mut client = Client::connect(DB_URL, NoTls)?;
 
-    //Create table
-    client.batch_execute(
-        "CREATE TABLE IF NOT EXISTS users (
-            id SERIAL PRIMARY KEY,
-            name VARCHAR NOT NULL,
-            email VARCHAR NOT NULL
-        )"
-    )?;
-    Ok(())
-}
+//     //Create table
+//     client.batch_execute(
+//         "CREATE TABLE IF NOT EXISTS users (
+//             id SERIAL PRIMARY KEY,
+//             name VARCHAR NOT NULL,
+//             email VARCHAR NOT NULL
+//         )"
+//     )?;
+//     Ok(())
+// }
 
-//get_id function
-fn get_id(request: &str) -> &str {
-    request.split("/").nth(2).unwrap_or_default().split_whitespace().next().unwrap_or_default()
-}
+// //get_id function
+// fn get_id(request: &str) -> &str {
+//     request.split("/").nth(2).unwrap_or_default().split_whitespace().next().unwrap_or_default()
+// }
 
-//deserialize user from request body with the id
-fn get_user_request_body(request: &str) -> Result<User, serde_json::Error> {
-    serde_json::from_str(request.split("\r\n\r\n").last().unwrap_or_default())
-}
+// //deserialize user from request body with the id
+// fn get_user_request_body(request: &str) -> Result<User, serde_json::Error> {
+//     serde_json::from_str(request.split("\r\n\r\n").last().unwrap_or_default())
+// }
