+++ src/models.rs
@@ -1,0 +1,7 @@
+//Model: User struct with id, name, email
+#[derive(Serialize, Deserialize)]
+pub struct User {
+    pub id: Option<i32>,
+    pub name: String,
+    pub email: String,
+}
