+++ src/utils.rs
@@ -1,0 +1,9 @@
+//get_id function
+pub fn get_id(request: &str) -> &str {
+    request.split("/").nth(2).unwrap_or_default().split_whitespace().next().unwrap_or_default()
+}
+
+//deserialize user from request body with the id
+pub fn get_user_request_body(request: &str) -> Result<crate::models::User, serde_json::Error> {
+    serde_json::from_str(request.split("\r\n\r\n").last().unwrap_or_default())
+}
